+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this library adheres to Rust's notion of
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `group::Wnaf` APIs for caching both bases and scalars, for improved many-base
+  many-scalar performance:
+  - `group::FixedWindow`
+  - `group::Wnaf::<FixedWindow<WINDOW_SIZE>, _, _>::{base, scalar, exp}`
 
 ## [0.12.0] - 2022-05-04
 ### Changed